requirements.txt
jsonrpclib-pelix
pika
pymongo
requests
redis
scikit-learn
numpy
scipy
python-dateutil
newspaper3k
tensorflow
pandas
watchdog